Revert "Superschemes for sketcher (#379)"
This reverts commit f7b31e7f653ea6e352aa60828b3ed1f6bea5c64e.
