+++ src/lib/forms/PipelineForm.svelte
@@ -6,37 +6,21 @@
     SecondaryButton,
     TextArea,
   } from "lib/govuk";
-  import Select from "lib/govuk/Select.svelte";
   import { prettyPrintMeters } from "lib/maplibre";
-  import { gjScheme, routeTool } from "stores";
+  import { routeTool } from "stores";
   import type { InterventionProps } from "types";
   import ATF4Type from "./ATF4Type.svelte";
 
   export let id: number;
   export let props: InterventionProps;
 
-  const mappedChoices: [string, string][] | undefined =
-    $gjScheme.subschemes?.map((subscheme) => {
-      return [subscheme.id.toString(), subscheme.name];
-    });
-
-  const schemeChoices: [string, string][] = mappedChoices ? mappedChoices : [];
-
   props.pipeline ||= {
-    schemeId: 0,
-    description: "",
-    type: "",
     atf4_type: "",
     accuracy: "",
     is_alternative: false,
   };
   // Guaranteed to exist
   let pipelineProps = props.pipeline;
-
-  let selectedSchemeIdString =
-    pipelineProps?.schemeId == 0 || pipelineProps?.schemeId
-      ? pipelineProps?.schemeId.toString()
-      : "";
 
   // Sets the intervention name to "From {road1 and road2} to {road3 and
   // road4}". Only meant to be useful for routes currently.
@@ -46,10 +30,6 @@
     } catch (e) {
       window.alert(`Couldn't auto-name route: ${e}`);
     }
-  }
-
-  function schemeSelectionChanged() {
-    pipelineProps.schemeId = parseInt(selectedSchemeIdString);
   }
 </script>
 
@@ -62,14 +42,6 @@
     </SecondaryButton>
   {/if}
 </FormElement>
-
-<Select
-  id="scheme-selector"
-  label="Scheme name:"
-  choices={schemeChoices}
-  bind:value={selectedSchemeIdString}
-  on:change={schemeSelectionChanged}
-/>
 
 <TextArea label="Description" bind:value={props.description} />
