+++ src/lib/sidebar/EntireScheme.svelte
@@ -10,6 +10,7 @@
   import { gjScheme, mode, sidebarHover } from "stores";
   import { onMount } from "svelte";
   import type { Schema, Scheme } from "types";
+  import PipelineSchemeForm from "./PipelineSchemeForm.svelte";
   import { backfill, interventionWarning } from "./scheme_data";
 
   export let authorityName: string;
@@ -155,6 +156,9 @@
         </SecondaryButton>
       </ButtonGroup>
     </Modal>
+    {#if schema == "pipeline"}
+      <PipelineSchemeForm />
+    {/if}
   </CollapsibleCard>
 
   {#if numErrors == 1}
