+++ src/lib/sidebar/PipelineSchemeForm.svelte
@@ -7,10 +7,20 @@
     SecondaryButton,
     TextArea,
   } from "lib/govuk";
-  import type { PipelineScheme } from "types";
+  import { gjScheme } from "stores";
   import ATF4Type from "../forms/ATF4Type.svelte";
 
-  export let subscheme: PipelineScheme;
+  $gjScheme.pipeline ||= {
+    scheme_type: "",
+    status: "",
+    timescale: "",
+    atf4_lead_type: "",
+    scheme_description: "",
+    funding_source: "",
+    funded: false,
+  };
+  // Guaranteed to exist
+  let pipelineProps = $gjScheme.pipeline;
 
   let showModal = false;
 
@@ -35,13 +45,13 @@
     ]}
     inlineSmall
     required
-    bind:value={subscheme.scheme_type}
+    bind:value={pipelineProps.scheme_type}
   />
 
   <ATF4Type
     label="Type of the main intervention"
     id="atf4-lead-type"
-    bind:value={subscheme.atf4_lead_type}
+    bind:value={pipelineProps.atf4_lead_type}
   />
 
   <fieldset class="govuk-fieldset">
@@ -58,7 +68,7 @@
       ]}
       inlineSmall
       required
-      bind:value={subscheme.status}
+      bind:value={pipelineProps.status}
     />
 
     <Radio
@@ -71,7 +81,7 @@
       ]}
       inlineSmall
       required
-      bind:value={subscheme.timescale}
+      bind:value={pipelineProps.timescale}
     />
     <div>
       <label>
@@ -80,7 +90,7 @@
           type="number"
           min="2023"
           max="2100"
-          bind:value={subscheme.timescale_year}
+          bind:value={pipelineProps.timescale_year}
         />
       </label>
     </div>
@@ -88,7 +98,7 @@
 
   <TextArea
     label="Scheme description"
-    bind:value={subscheme.scheme_description}
+    bind:value={pipelineProps.scheme_description}
   />
 
   <fieldset class="govuk-fieldset">
@@ -97,13 +107,17 @@
     <div>
       <label>
         GBP funded
-        <input type="number" min="0" bind:value={subscheme.budget_funded} />
+        <input type="number" min="0" bind:value={pipelineProps.budget_funded} />
       </label>
     </div>
     <div>
       <label>
         GBP unfunded
-        <input type="number" min="0" bind:value={subscheme.budget_unfunded} />
+        <input
+          type="number"
+          min="0"
+          bind:value={pipelineProps.budget_unfunded}
+        />
       </label>
     </div>
 
@@ -119,10 +133,10 @@
         repeat("LUF"),
       ]}
       inlineSmall
-      bind:value={subscheme.funding_source}
+      bind:value={pipelineProps.funding_source}
     />
 
-    <Checkbox id="scheme-funded" bind:checked={subscheme.funded}>
+    <Checkbox id="scheme-funded" bind:checked={pipelineProps.funded}>
       Is the scheme fully funded?
     </Checkbox>
   </fieldset>
@@ -134,7 +148,7 @@
         type="number"
         min="2010"
         max="2100"
-        bind:value={subscheme.source_data_year}
+        bind:value={pipelineProps.source_data_year}
       />
     </label>
   </div>
