+++ tests/shared.ts
@@ -3,24 +3,8 @@ import { expect, type Browser, type Page } from "@playwright/test";
 export async function loadInitialPageFromBrowser(
   browser: Browser
 ): Promise<Page> {
-  return loadPageFromBrowser(browser, "/scheme.html?authority=LAD_Adur");
-}
-
-export async function loadPipelineSketchPageFromBrowser(
-  browser: Browser
-): Promise<Page> {
-  return loadPageFromBrowser(
-    browser,
-    "/scheme.html?authority=LAD_Adur&schema=pipeline"
-  );
-}
-
-export async function loadPageFromBrowser(
-  browser: Browser,
-  url: string
-): Promise<Page> {
   let page = await browser.newPage();
-  await page.goto(url);
+  await page.goto("/scheme.html?authority=LAD_Adur");
   await checkPageLoaded(page);
   return page;
 }
